Bookshelf App
This website helps users keep track of a personal bookshelf so they can manage books they are reading.  This is my final project for Code Louisville's front end web dev 2 class. The book search feature uses the Google Books API, and by clicking on the "add to bookshelf" button the book is saved to local storage.  This site is responsive on different devices through the use of media queries.  

The three features used are:
1.) Retrieve data from an external API and display data in your app (such as with fetch() or with AJAX)
2.) Create an array, dictionary or list, populate it with multiple values, retrieve at least one value, and use or display it in your application.
3.) Create a form and save the values (on click of Submit button) to an external file 
You must show us or document where that information is being stored so we can confirm that it’s being saved/persisted. 
*(In browser- right click ->Inspect -> Application -> Local Storage)*